Staging

+++ src/main/java/id/ac/ui/cs/advprog/buy/controller/BuyController.java
@@ -1,11 +1,14 @@
 package id.ac.ui.cs.advprog.buy.controller;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import id.ac.ui.cs.advprog.buy.middleware.AuthMiddleware;
 import id.ac.ui.cs.advprog.buy.model.Cart;
 import id.ac.ui.cs.advprog.buy.model.Transaction;
 import id.ac.ui.cs.advprog.buy.model.TransactionFactory;
 import id.ac.ui.cs.advprog.buy.service.CartService;
 import id.ac.ui.cs.advprog.buy.service.TransactionService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.configurationprocessor.json.JSONException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -24,6 +27,7 @@ public class BuyController {
 
     @Autowired
     private TransactionService transactionService;
+
     @GetMapping("/")
     @ResponseBody
     public String buyPage(){
@@ -31,50 +35,92 @@ public class BuyController {
     }
 
     @PostMapping("/cart")
-    public ResponseEntity<Cart> addToCart(@RequestParam("usr") String username, @RequestBody Map<String,Integer> addListings){
+    public ResponseEntity<?> addToCart(@RequestHeader("Authorization") String token, @RequestBody Map<String,Integer> addListings){
+        String username = AuthMiddleware.getUsernameFromToken(token);
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        ResponseEntity<?> authResponse = authenticateBuyerSeller(username,userRole);
+        if (authResponse != null) {
+            return authResponse;
+        }
+
+        Cart cart;
         try {
-            Cart cart = cartService.addListings(addListings,username);
-            // TODO : Jalankan service update price
-            return ResponseEntity.ok(cart);
+            cart = cartService.addListings(addListings,username);
         } catch (NoSuchElementException e){
-            Cart cart = new Cart(username);
+            cart = new Cart(username);
             cart.setListings(addListings);
             cartService.create(cart);
-            // TODO : Jalankan service update price
+        }
+
+        try {
+            cartService.updateTotalPrice(username,token);
             return ResponseEntity.ok(cart);
+        } catch (JSONException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to parse JSON");
         }
     }
 
     @PutMapping("/cart/reduce")
-    public ResponseEntity<?> reduceListing(@RequestParam("usr") String username, @RequestBody Map<String,Integer> reduceListings){
+    public ResponseEntity<?> reduceListing(@RequestHeader("Authorization") String token, @RequestBody Map<String,Integer> reduceListings){
+        String username = AuthMiddleware.getUsernameFromToken(token);
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        ResponseEntity<?> authResponse = authenticateBuyerSeller(username,userRole);
+        if (authResponse != null) {
+            return authResponse;
+        }
+
         try{
             Cart updatedCart = cartService.reduceListings(reduceListings,username);
-            // TODO : Jalankan update price
+            cartService.updateTotalPrice(username,token);
             return ResponseEntity.ok(updatedCart);
         } catch (NoSuchElementException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cart not found for user: " + username);
+        } catch (JSONException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to parse JSON");
         }
     }
 
     @PutMapping("/cart/add")
-    public ResponseEntity<?> addListing(@RequestParam("usr") String username, @RequestBody Map<String,Integer> addListings){
+    public ResponseEntity<?> addListing(@RequestHeader("Authorization") String token, @RequestBody Map<String,Integer> addListings){
+        String username = AuthMiddleware.getUsernameFromToken(token);
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        ResponseEntity<?> authResponse = authenticateBuyerSeller(username,userRole);
+        if (authResponse != null) {
+            return authResponse;
+        }
+
         try{
             Cart updatedCart = cartService.addListings(addListings,username);
-            // TODO : Jalankan update price
+            cartService.updateTotalPrice(username,token);
             return ResponseEntity.ok(updatedCart);
         } catch (NoSuchElementException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cart not found for user: " + username);
+        } catch (JSONException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to parse JSON");
         }
     }
 
     @DeleteMapping("/cart")
-    public ResponseEntity<?> deleteListing(@RequestParam("usr")String username, @RequestParam("lstId") String listingId){
+    public ResponseEntity<?> deleteListing(@RequestParam("lstId") String listingId, @RequestHeader("Authorization") String token){
+        String username = AuthMiddleware.getUsernameFromToken(token);
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        ResponseEntity<?> authResponse = authenticateBuyerSeller(username,userRole);
+        if (authResponse != null) {
+            return authResponse;
+        }
+
         try{
             Cart updatedCart = cartService.deleteListing(listingId,username);
-            // TODO : Jalankan update price
+            cartService.updateTotalPrice(username,token);
             return ResponseEntity.ok(updatedCart);
         } catch (NoSuchElementException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cart not found for user: " + username);
+        } catch (JSONException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to parse JSON");
         }
     }
 
@@ -95,11 +141,18 @@ public class BuyController {
     }
 
     @PostMapping("/transaction")
-    public ResponseEntity<?> checkout(@RequestBody Map<String,String> userData){
-        String username = userData.get("username");
+    public ResponseEntity<?> checkout(@RequestBody Map<String,String> userData, @RequestHeader("Authorization") String token){
+        String username = AuthMiddleware.getUsernameFromToken(token);
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        ResponseEntity<?> authResponse = authenticateBuyerSeller(username,userRole);
+        if (authResponse != null) {
+            return authResponse;
+        }
+
         String deliveryLocation = userData.get("deliveryLocation");
 
-        if (username == null || deliveryLocation == null){
+        if (deliveryLocation == null){
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid user data");
         }
 
@@ -108,16 +161,28 @@ public class BuyController {
             UUID uuid = UUID.randomUUID();
 
             Transaction transaction = TransactionFactory.createTransaction(selectedCart,uuid.toString(),deliveryLocation);
-            transactionService.create(transaction);
+            transactionService.create(transaction,token);
             return ResponseEntity.ok(transaction);
 
         } catch (NoSuchElementException e){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cart not found for user: " + username);
+        } catch (JsonProcessingException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Could not create payment transaction");
         }
     }
 
     @PutMapping("/transaction/{transactionid}")
-    public ResponseEntity<?> updateTransactionStatus(@PathVariable String transactionid, @RequestParam("status") String status){
+    public ResponseEntity<?> updateTransactionStatus(@PathVariable String transactionid, @RequestParam("status") String status, @RequestHeader("Authorization") String token){
+        String userRole = AuthMiddleware.getRoleFromToken(token);
+
+        if (userRole == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+
+        if (!userRole.equals("STAFF")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Role must be Staff");
+        }
+
         try{
             Transaction transaction = transactionService.updateStatus(transactionid,status);
             return ResponseEntity.ok(transaction);
@@ -148,6 +213,14 @@ public class BuyController {
         }
     }
 
+    private ResponseEntity<?> authenticateBuyerSeller(String username,String userRole) {
+        if (username == null || userRole == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
+        }
+        if (userRole.equals("STAFF")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Role must be Buyer/Seller");
+        }
 
-
+        return null;
+    }
 }
